requirements.txt
python-telegram-bot
python-dotenv
google-generativeai
google-genai
PyPDF2
